requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-SocketIO==5.3.6
python-dotenv==1.0.1
psycopg2-binary==2.9.9
pandas==2.2.3
numpy==2.1.1
eventlet==0.36.1
